Day 10 puzzle 2
Semi-manual process. Could have been automated more, but it was quicker
to separate out the different lists using a spreadsheet.
The list was split whenever there were two consecutive diffs of 3. For
simplicity the data was then adjusted in the sublist to treat the first
item as the 0 starting point.